TsundereBase.class
Made a base class
normal tsundereclass renamed to TsundereEnglish -> not ingame tho
